# Conflicts:
#BCH/library/api/fermat-bch-api/build.gradle
#CHT/android/sub_app/fermat-cht-android-sub-app-chat-identity-bitdubai/src/main/res/layout/fragment_cht_identity_geolocation.xml
#OSA/settings.gradle
#WPD/plugin/engine/fermat-wpd-plugin-engine-app-runtime-bitdubai/src/main/java/com/bitdubai/fermat_wpd_plugin/layer/engine/wallet_runtime/developer/bitdubai/version_1/AppRuntimeEnginePluginRoot.java
#fermat-api/src/main/java/com/bitdubai/fermat_api/layer/all_definition/navigation_structure/enums/Activities.java
#fermat-core/src/main/java/com/bitdubai/fermat_core/FermatSystem.java
update with content from android-upgrade

+++ CHT/library/api/fermat-cht-api/src/main/java/com/bitdubai/fermat_cht_api/layer/sup_app_module/interfaces/chat_actor_community/interfaces/Cities.java
@@ -36,13 +36,13 @@ public interface Cities extends Serializable {
      * This method returns the city latitude.
      * @return
      */
-    float getLatitude();
+    double getLatitude();
 
     /**
      * This method returns the city longitude.
      * @return
      */
-    float getLongitude();
+    double getLongitude();
 
     /**
      * This method returns the country code from a city.
